+++ csatar-plugins/csatar/Plugin.php
@@ -183,6 +183,7 @@ class Plugin extends PluginBase
             \Csatar\Csatar\Components\OrganizationUnitFrontend::class => 'organizationUnitFrontend',
             \Csatar\Csatar\Components\CsatarGallery::class => 'csatargallery',
             \Csatar\Csatar\Components\Breadcrumb::class => 'csatarBreadcrumb',
+            \Csatar\Csatar\Components\TwoFactorAuthentication::class => 'twoFactorAuthentication',
         ];
     }