+++ csatar-plugins/csatar/components/TwoFactorAuthentication.php
@@ -1,0 +1,70 @@
+<?php namespace Csatar\Csatar\Components;
+
+use Auth;
+use Cms\Classes\ComponentBase;
+use Csatar\Csatar\Classes\GoogleTwoFactorAuthentication;
+use Input;
+use Lang;
+use Session;
+
+class TwoFactorAuthentication extends ComponentBase
+{
+    private $userSecretKey;
+    private $google2FA;
+    public $qrCodeData;
+    public $is2FAuthenticated;
+
+    public function componentDetails()
+    {
+        return [
+            'name' => Lang::get('csatar.csatar::lang.plugin.component.twoFactorAuthentication.name'),
+            'description' => Lang::get('csatar.csatar::lang.plugin.component.twoFactorAuthentication.description'),
+        ];
+    }
+
+    public function onRun()
+    {
+        if (!Auth::user()) {
+            return;
+        }
+
+        if (!isset(Auth::user()->scout)) {
+            return;
+        }
+
+        $this->prepareVariables();
+    }
+
+    public function onVerifyCode()
+    {
+        $this->prepareVariables();
+
+        if ($this->google2FA->verifyKey($this->userSecretKey, Input::get('code'))) {
+            Session::put('scout.twoFA', true);
+            return [
+                '#qrInputContainer' => $this->renderPartial('@success')
+            ];
+        } else {
+            throw new \ValidationException(['code' => Lang::get('csatar.csatar::lang.plugin.component.twoFactorAuthentication.twoFactorAuthFailed')]);
+        }
+    }
+
+    private function prepareVariables() {
+        $scout = Auth::user()->scout;
+        $this->google2FA = new GoogleTwoFactorAuthentication();
+
+        if (empty($scout->google_two_fa_secret_key)) {
+            $scout->google_two_fa_secret_key = $this->google2FA->generateSecretKey();
+            $scout->forceSave();
+        }
+
+        $this->userSecretKey = $scout->google_two_fa_secret_key;
+        $this->qrCodeData = $this->google2FA->getQRCodeData('RMCSSZ', $scout->getFullName(), $this->userSecretKey, 300);
+        $this->is2FAuthenticated = $this->is2FAuthenticated();
+    }
+
+    private function is2FAuthenticated()
+    {
+        return Session::get('scout.twoFA') == true;
+    }
+}