+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- CS-36 - Implemented two-factor authentication 
 - CS-338 - Reduced EcsetCode length to 6 characters + separator + suffix
 - CS-340 - Hotfix - Changed preselected disabled fields to readonly
 - CS-302 - Modified navigation bar to support multi-level